Initial Code
The file scans through each team folder and creates a team dictionary with the key being the player name and the values being the player data from the excel sheets.

There is an algorithm to count the hit rate of each player prop with the value being 90% of there average. It then prints the result