Terminal menu added
There is an interactable menu which allows the user to select specific players  and their props. It first selects the team, then player, then stat, and takes a user input for a line to compare. It would then print the hit rate for said line.

There is some repeating code which I can put into its own function that I will address. As well as the try exception lines printing an additional output