Created prop dictionary
A dictionary is created that contains a data frame of each players prop lines with the key being their name. This is to minimize the amount of calls to the API for the game logs. It only has to get it once for the player and then analyzes it with the prop dictionary